feat: Add new movie form with Material UI inputs and custom styles
Implemented a form for adding a new movie, using Material UI components such as TextField, Select, and Fab. Custom styles were created using styled-components for the Material UI components, in order to adjust the appearance of the inputs and the button for adding a new category. PropTypes were also added for prop validation.

Changes made:
- Added new movie form with Material UI inputs
- Implemented custom styles for the inputs using styled-components
- Added PropTypes for prop validation